+++ examples/binding_test.rs
@@ -67,12 +67,12 @@ fn main() {
     #[cfg(feature = "rust-alloc")]
     lv_bevy_ecs::malloc::set_mem_monitor(get_memory_stats);
 
-    const HOR_RES: u32 = 800;
-    const VER_RES: u32 = 480;
-    const LINE_HEIGHT: u32 = 10;
+    const HOR_RES: usize = 800;
+    const VER_RES: usize = 480;
+    const LINE_HEIGHT: usize = 10;
 
     let mut sim_display: SimulatorDisplay<Rgb565> =
-        SimulatorDisplay::new(Size::new(HOR_RES, VER_RES));
+        SimulatorDisplay::new(Size::new(HOR_RES as u32, VER_RES as u32));
 
     let output_settings = OutputSettingsBuilder::new().scale(1).build();
     let mut window = Window::new("Bindings Test Example", &output_settings);
